Split up AssetCard
